requirements.txt
requests==2.28.1
python-dotenv==0.21.0
googlemaps==4.6.0
Django==3.2.16